[taken from gcc-2.95.3-latest-cvs] 	fix incorrect type of sz
git-svn-id: file:///srv/svn/repos/haiku/trunk/buildtools@9598 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ gcc/libio/SFile.h
@@ -35,7 +35,7 @@ extern "C++" {
 class SFile: public fstream
 {
   protected:
-    int       sz;                   // unit size for structured binary IO
+    long      sz;                   // unit size for structured binary IO
 
 public:
     SFile() : fstream() { }
